ha2: fixed based on review
