+++ ha2/CMakeLists.txt
@@ -6,6 +6,6 @@ cmake_minimum_required (VERSION 3.8)
 project ("ha2")
 
 # Add source to this project's executable.
-add_executable (ha2 "ha2.cpp"  "templates.h" "count.h")
+add_executable (ha2 "ha2.cpp"  "templates.h" "count.h" "count.cpp")
 
 # TODO: Add tests and install targets if needed.
